Use the picker to select article target in news and calendar (see #6551)
Description
-----------

News and calendar events currently allow to select an article target. They render a dropdown list of all articles of visible root nodes.

With this PR, the user can now use the picker to select an article, which I think is a superior interface. The only disadvantage is that one must have access to articles in the backend. We _could_ use an callback to modify the field on-the-fly, but I'm not sure that's worth the effort?

PS: I remember having discussed this a long time ago and I was surprised this isn't implemented yet. Maybe someone remembers if I'm missing something?

Commits
-------

ed2f8227 Use the picker to select article target in news and calendar
47ea4c1e Also remove tl_calendar_events::getArticleAlias