Improve the `CalendarFeedListenerTest` (see #9088)
Description
-----------

As noted in https://github.com/contao/contao/pull/9087#discussion_r2598906150 the current `CalendarFeedListenerTest::testTransformsEventsToFeedItems()` test is flawed, as it does not actually distinguish between `source_teaser` and `source_text` properly. This PR fixes that.

The same applies to the `NewsFeedListenerTest`, but that will be a separate PR for `5.3`.

Commits
-------

3911ee4d Fix the CalendarFeedListenerTest